bind getWallets like service
